+++ analysis/src/SemanticTokens.ml
@@ -265,7 +265,9 @@ let command ~debug ~emitter ~path =
       *)
       emitter (* --> <div... *)
       |> emitJsxTag ~debug ~name:"<" ~pos:(Loc.start e.pexp_loc);
-      emitter |> emitJsxOpen ~lid:lident.txt ~debug ~loc:lident.loc;
+      let lid = Ast_helper.Jsx.longident_of_jsx_tag_name lident.txt in
+      let loc = lident.loc in
+      emitter |> emitJsxOpen ~lid ~debug ~loc;
       let closing_line, closing_column = Loc.end_ e.pexp_loc in
       emitter (* <foo ...props /> <-- *)
       |> emitJsxTag ~debug ~name:"/>" ~pos:(closing_line, closing_column - 2)
@@ -281,7 +283,9 @@ let command ~debug ~emitter ~path =
       (* opening tag *)
       emitter (* --> <div... *)
       |> emitJsxTag ~debug ~name:"<" ~pos:(Loc.start e.pexp_loc);
-      emitter |> emitJsxOpen ~lid:lident.txt ~debug ~loc:lident.loc;
+      let lid = Ast_helper.Jsx.longident_of_jsx_tag_name lident.txt in
+      let loc = lident.loc in
+      emitter |> emitJsxOpen ~lid ~debug ~loc;
       emitter (* <foo ...props > <-- *)
       |> emitJsxTag ~debug ~name:">"
            ~pos:(Pos.ofLexing posOfGreatherthanAfterProps);
@@ -308,9 +312,11 @@ let command ~debug ~emitter ~path =
              emitter
              |> emitJsxTag ~debug ~name:"</"
                   ~pos:(Pos.ofLexing closing_less_than);
-             emitter
-             |> emitJsxClose ~debug ~lid:lident.txt
-                  ~pos:(Loc.start tag_name_end.loc);
+             let lid =
+               Ast_helper.Jsx.longident_of_jsx_tag_name tag_name_end.txt
+             in
+             let loc = tag_name_end.loc in
+             emitter |> emitJsxClose ~debug ~lid ~pos:(Loc.end_ loc);
              emitter (* <foo> ... </foo> <-- *)
              |> emitJsxTag ~debug ~name:">"
                   ~pos:(Pos.ofLexing final_greather_than))