+++ compiler/ml/ast_helper.mli
@@ -212,13 +212,13 @@ module Exp : sig
   val jsx_unary_element :
     ?loc:loc ->
     ?attrs:attrs ->
-    Longident.t Location.loc ->
+    Parsetree.jsx_tag_name Location.loc ->
     Parsetree.jsx_props ->
     expression
   val jsx_container_element :
     ?loc:loc ->
     ?attrs:attrs ->
-    Longident.t Location.loc ->
+    Parsetree.jsx_tag_name Location.loc ->
     Parsetree.jsx_props ->
     Lexing.position ->
     Parsetree.jsx_children ->
@@ -299,6 +299,11 @@ module Te : sig
     str ->
     extension_constructor
   val rebind : ?loc:loc -> ?attrs:attrs -> str -> lid -> extension_constructor
+end
+
+module Jsx : sig
+  val string_of_jsx_tag_name : Parsetree.jsx_tag_name -> string
+  val longident_of_jsx_tag_name : Parsetree.jsx_tag_name -> Longident.t
 end
 
 (** {1 Module language} *)