+++ compiler/ml/location.ml
@@ -300,3 +300,5 @@ let raise_errorf ?(loc = none) ?(sub = []) ?(if_highlight = "") =
 
 let deprecated ?(def = none) ?(use = none) loc msg =
   prerr_warning loc (Warnings.Deprecated (msg, def, use))
+
+let map_loc f {txt; loc} = {txt = f txt; loc}